requirements.txt
fastapi~=0.112.2
pydantic~=2.8.2
starlette~=0.38.2
SQLAlchemy~=2.0.32
jose~=1.0.0
passlib~=1.7.4
python-jose~=3.3.0
azure-core~=1.30.2
python-dotenv~=1.0.1